+++ routes/imports.js
@@ -124,23 +124,6 @@ exports.render = function(req, res) {
     })
 }
 
-// GET request to the /google page (view settings)
-
-exports.renderGoogle = function(req, res) {
-    var contextslist = []
-    if (res.locals.contextslist) {
-        contextslist = res.locals.contextslist
-    }
-    res.render('google', {
-        title: 'Google the Google',
-        evernote: '',
-        context: req.query.context,
-        notebooks: '',
-        contextlist: contextslist,
-        fornode: req.query.fornode,
-    })
-}
-
 exports.renderYouTube = function(req, res) {
     var contextslist = []
     if (res.locals.contextslist) {
@@ -2623,271 +2606,8 @@ exports.submit = function(req, res, next) {
                 }
             }
         })
-    } else if (service == 'googlesearch') {
-        // google.lang = 'us'
-        // google.tld = 'us'
+    } 
 
-        // google.resultsPerPage = limit
-        // var nextCounter = 0
-
-        var statements = []
-
-        var default_context = importContext
-
-        var addToContexts = []
-        addToContexts.push(default_context)
-
-        var excludesearchquery = req.body.excludesearchquery
-
-        var excludetitles = req.body.excludetitles
-
-
-        validate.getContextID(user_id, addToContexts, function(result, err) {
-            if (err) {
-                console.log(err)
-                res.error(
-                    'Something went wrong when adding a new context into the database. Try changing its name or open an issue on GitHub.'
-                )
-                res.redirect('back')
-            } else {
-                var contexts = result
-                var req = {
-                    body: {
-                        entry: {
-                            body: [],
-                        },
-                        context: default_context,
-                    },
-
-                    contextids: contexts,
-                    internal: 1,
-                    multiple: 1,
-                    walkthrough: walknext
-                }
-
-
-                let google_request_link = config.google.URL_search + config.google.API_key + '&q=' + searchString.toLowerCase();
-
-                https.get(google_request_link, (resp) => {
-
-                    var receiveddata = '';
-
-                    // A chunk of data has been received.
-                    resp.on('data', (chunk) => {
-                        receiveddata += chunk;
-                    });
-
-                    // The whole response has been received. Print out the result.
-                    resp.on('end', () => {
-
-                        let googlejson = JSON.parse(receiveddata);
-
-                        req.body.entry.body = addGoogleEntry(googlejson, excludetitles);
-
-                        let currentmarker = 0;
-
-                        // How much total results we get?
-                        let tot_search_results = googlejson.queries.request[0].totalResults;
-
-                        currentmarker = tot_search_results - 10;
-
-                        let startmarker = 0;
-
-                        if (currentmarker > 0) {
-                            startmarker = 10;
-                            https.get(google_request_link + '&start=' + startmarker, (resp) => {
-
-                                receiveddata = '';
-                                // A chunk of data has been received.
-                                resp.on('data', (chunk) => {
-                                    receiveddata += chunk;
-                                });
-
-                                // The whole response has been received. Print out the result.
-                                resp.on('end', () => {
-
-                                    googlejson = JSON.parse(receiveddata);
-
-                                    req.body.entry.body = req.body.entry.body.concat(addGoogleEntry(googlejson, excludetitles));
-
-                                    currentmarker = tot_search_results - 20;
-
-                                    if (currentmarker > 0) {
-                                        startmarker = 20;
-
-                                        https.get(google_request_link + '&start=' + startmarker, (resp) => {
-
-                                            receiveddata = '';
-                                            // A chunk of data has been received.
-                                            resp.on('data', (chunk) => {
-                                                receiveddata += chunk;
-                                            });
-
-                                            // The whole response has been received. Print out the result.
-                                            resp.on('end', () => {
-
-                                                googlejson = JSON.parse(receiveddata);
-
-                                                req.body.entry.body = req.body.entry.body.concat(addGoogleEntry(googlejson, excludetitles));
-
-                                                currentmarker = tot_search_results - 30;
-
-                                                if (currentmarker > 0) {
-                                                    startmarker = 30;
-
-                                                    https.get(google_request_link + '&start=' + startmarker, (resp) => {
-
-                                                        receiveddata = '';
-                                                        // A chunk of data has been received.
-                                                        resp.on('data', (chunk) => {
-                                                            receiveddata += chunk;
-                                                        });
-
-                                                        // The whole response has been received. Print out the result.
-                                                        resp.on('end', () => {
-
-                                                            googlejson = JSON.parse(receiveddata);
-
-                                                            req.body.entry.body = req.body.entry.body.concat(addGoogleEntry(googlejson, excludetitles));
-
-
-                                                             // Do we need to exclude the search query from the graph? Let's make a temporary list of stopwords for it
-
-                                                            if (excludesearchquery) {
-
-                                                                let searchterms = searchString.toLowerCase().split(' ')
-
-                                                                let searchlemmas = [];
-
-                                                                for (let k = 0; k < searchterms.length; k++) {
-
-                                                                    // Now we find lemmas, so we deal with plural cases and also with Russian word endings and suffixes
-                                                                    // TODO this whole thing should be moved outside of this function and Russian lemmas should be added to stopwords not deleted from text
-
-                                                                    if (
-                                                                        /[а-яА-ЯЁё]/.test(searchterms[k]) ==
-                                                                        true
-                                                                    ) {
-                                                                        var lemmaterm = lemmerRus.lemmatize(
-                                                                            searchterms[k]
-                                                                        )
-                                                                    }
-
-                                                                    // English?
-                                                                    else {
-                                                                        var lemmaterm = lemmerEng.lemmatize(
-                                                                            searchterms[k]
-                                                                        )
-
-                                                                    }
-
-                                                                    // Now push the lemma into the list
-                                                                    if (lemmaterm[0] != undefined) {
-                                                                        searchlemmas.push(lemmaterm[0].toLowerCase())
-                                                                    }
-                                                                }
-
-                                                                req.excludestopwords = searchlemmas;
-                                                            }
-
-                                                            // Submit entries (routes/entries.js)
-
-                                                            entries.submit(req, res);
-
-
-                                                        });
-                                                    });
-
-                                                }
-                                                else {
-                                                    entries.submit(req, res);
-
-                                                }
-
-
-                                            });
-                                        });
-
-                                    }
-                                    else {
-                                        entries.submit(req, res);
-                                    }
-
-
-                                });
-                            });
-                        }
-                        else {
-                            if (tot_search_results > 0) {
-                                entries.submit(req, res);
-                            }
-                            else {
-                                console.log("Error: No Google search results found");
-                                res.error(JSON.stringify("Google provided zero search results for this query. Try to make it a bit more general."));
-                                res.redirect('back')
-                            }
-                        }
-
-
-
-
-
-                    });
-
-
-
-                }).on("error", (err) => {
-                    console.log("Error: " + err.message);
-                    res.error(JSON.stringify(err.message));
-                    res.redirect('back')
-                });
-
-            }
-        })
-    }
-
-
-    function addGoogleEntry(googlejson, excludetitles) {
-
-        console.log('googlejson')
-            console.log(googlejson)
-
-            let entryobject = [];
-
-            let searchresults = googlejson.items;
-
-            if (searchresults) {
-                for (let i = 0; i < searchresults.length;  ++i) {
-                    if (
-                        searchresults[i].snippet &&
-                        searchresults[i].snippet != 'null' &&
-                        searchresults[i].snippet != undefined
-                    ) {
-
-                        let searchtext = ''
-
-                        if (!excludetitles) {
-                            searchtext += searchresults[i].title + ' ';
-                        }
-                        searchtext += searchresults[i].snippet
-                        searchtext += ' ' + searchresults[i].link
-
-                        // Old Google import code that replaces the dates
-                        // searchtext = searchtext.replace(
-                        //     /(0?[1-9]|[12][0-9]|3[01])\s{1}(Jan|Feb|Mar|Apr|May|Jun|Jul|Apr|Sep|Oct|Nov|Dec)\s{1}\d{4}/g,
-                        //     ''
-                        // )
-
-                        // Add the search result as an entry to process
-                        entryobject[i] = searchtext;
-
-                    }
-                }
-            }
-
-            return entryobject;
-
-    }
 
     function saveHighlight(highlight, contexts) {
         // and finally create an object to send this entry with the right context
