+++ views/google.ejs
@@ -33,7 +33,7 @@
         <div id='welcomemessage'>
 
 
-            <form action='/import' name='submitform' id="submitform" method='post' class='pure-form pure-form-stacked'>
+            <form action='/importgoogle' name='submitform' id="submitform" method='post' class='pure-form pure-form-stacked'>
                 Type in your search query and visualize the available Google search results as a graph, so you can see the main topics, how they relate, what's there and what's missing.
                 <br>&nbsp;<br>
                 <input type='text' size="25" maxlength="50" name='search' id="search" class="pure-input" placeholder='search terms' value="<% if (fornode) { %><%= fornode %><% } %>" /> <br><br><input class="pure-button pure-button-primary" type='submit' name="btnSubmit" value="Import">
